requirements.txt
Django==1.8
djangorestframework==3.3.3
httpie==0.9.3
requests==2.9.1
django-rest-swagger==0.3.5